+++ system/src/Grav/Framework/File/Formatter/YamlFormatter.php
@@ -30,11 +30,20 @@ class YamlFormatter implements FormatterInterface
     }
 
     /**
-     * {@inheritdoc}
+     * @deprecated
+     * @internal
      */
     public function getFileExtension()
     {
-        return $this->config['file_extension'];
+        return $this->getSupportedFileExtensions()[0];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getSupportedFileExtensions()
+    {
+        return (array) $this->config['file_extension'];
     }
 
     /**